(#28)  Split `commands.py` into the interface and the implementations (the new `scrapers.py`).
